Use int64 for uid and gid everywhere
The right type on Linux would be uint32 but a number of functions rely
on negative values so int64 will be easier to deal with.

+++ lxd/migrate.go
@@ -310,9 +310,9 @@ func (s *migrationSourceWs) Do(migrateOp *operation) error {
 			idmap := IDMapType{
 				Isuid:    proto.Bool(ctnIdmap.Isuid),
 				Isgid:    proto.Bool(ctnIdmap.Isgid),
-				Hostid:   proto.Int(ctnIdmap.Hostid),
-				Nsid:     proto.Int(ctnIdmap.Nsid),
-				Maprange: proto.Int(ctnIdmap.Maprange),
+				Hostid:   proto.Int32(int32(ctnIdmap.Hostid)),
+				Nsid:     proto.Int32(int32(ctnIdmap.Nsid)),
+				Maprange: proto.Int32(int32(ctnIdmap.Maprange)),
 			}
 
 			idmaps = append(idmaps, &idmap)
@@ -760,9 +760,9 @@ func (c *migrationSink) Do(migrateOp *operation) error {
 			e := shared.IdmapEntry{
 				Isuid:    *idmap.Isuid,
 				Isgid:    *idmap.Isgid,
-				Nsid:     int(*idmap.Nsid),
-				Hostid:   int(*idmap.Hostid),
-				Maprange: int(*idmap.Maprange)}
+				Nsid:     int64(*idmap.Nsid),
+				Hostid:   int64(*idmap.Hostid),
+				Maprange: int64(*idmap.Maprange)}
 			srcIdmap.Idmap = shared.Extend(srcIdmap.Idmap, e)
 		}
